features: add the echo command
