--- src/main/java/Duke.java
@@ -1,13 +1,0 @@
-public class Duke {
-
-  public static void main(String[] args) {
-    String welcomeMessage =
-      "____________________________________________________________\n" +
-      "Hello! I'm ATSISBot\n" +
-      "What can I do for you?\n" +
-      "____________________________________________________________\n" +
-      "Bye. Hope to see you again soon!\n" +
-      "____________________________________________________________\n";
-    System.out.println(welcomeMessage);
-  }
-}
